refactor(ts): cleanup methods of `RolldownBuild`
