+++ packages/node/src/utils/index.ts
@@ -2,6 +2,7 @@ export * from './async-flatten'
 export * from './transform-to-rollup-output'
 export * from './normalize-plugin-option'
 export * from './ensure-array'
+export * from './create-bundler'
 
 export function arraify<T>(value: T | T[]): T[] {
   return Array.isArray(value) ? value : [value]
